+++ rules/user_mention.test.js
@@ -54,10 +54,25 @@ it('handles a mention with text before (double new line)', () => {
 `);
 });
 
+it('does not inject mentions inside links', () => {
+    expect(md.render('[test@test](https://test.com)')).toBe(`<p><a href="https://test.com">test@test</a></p>
+`);
+});
+
+it('handles a mention that contains markdown', () => {
+    expect(md.render('**@test**')).toBe(`<p>**<a href="/users/test**">@test**</a></p>
+`);
+});
+
 const mdPattern = require('markdown-it')().use(require('./user_mention'), { pattern: /[a-z]+/i });
 
 it('handles a mention using a specific pattern', () => {
     expect(mdPattern.render('hello @test, thanks')).toBe(`<p>hello <a href="/users/test">@test</a>, thanks</p>
+`);
+});
+
+it('handles a mention inside other markdown when using a specific pattern', () => {
+    expect(mdPattern.render('hello **@test**, thanks')).toBe(`<p>hello <strong><a href="/users/test">@test</a></strong>, thanks</p>
 `);
 });
 
@@ -68,10 +83,15 @@ const mdPath = require('markdown-it')().use(require('./user_mention'), {
      * @param {string} mention User mention to generate a URL path for.
      * @returns {string}
      */
-    path: mention => `/world/${mention}`,
+    path: mention => (mention === 'blank' ? '' : `/world/${mention}`),
 });
 
 it('handles a mention using a specific link path', () => {
     expect(mdPath.render('hello @test')).toBe(`<p>hello <a href="/world/test">@test</a></p>
+`);
+});
+
+it('does not inject mentions when there is no path', () => {
+    expect(mdPath.render('hello @blank')).toBe(`<p>hello @blank</p>
 `);
 });